# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Gadget Inventory API
  version: 1.0.0
  description: API documentation for managing gadgets in the inventory. Includes CRUD operations for gadgets, with additional self-destruct functionality.
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:  
      type: http
      scheme: bearer
      bearerFormat: JWT 
      description: Use the Authorization header with the Bearer token to authenticate.
paths:
  /api/user/signup:
    post:
      summary: User registration (sign up)
      description: Allows users to register by providing a username and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: johndoe
                password:
                  type: string
                  example: password123
                type:
                  type: string
                  enum:
                    - admin
                    - user
                  example: user
      responses:
        200:
          description: Successfully created user
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                    example: 1
        400:
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid Entry"
        500:
          description: Internal server error
  /api/user/signin:
    post:
      summary: User login (sign in)
      description: Allows users to login by providing a username and password. Returns a JWT token upon successful authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: johndoe
                password:
                  type: string
                  example: password123
      responses:
        200:
          description: Successfully authenticated, returns a JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "your-jwt-token-here"
        404:
          description: User not found or incorrect password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        400:
          description: Invalid input or internal error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Incorrect Details"
...